+++ examples/standalone/expo-todomvc/metro.config.js
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-require-imports */
 /* eslint-disable unicorn/prefer-module */
 
 // Learn more https://docs.expo.io/guides/customizing-metro