+++ packages/@livestore/devtools-node-common/src/devtools-bridge/index.ts
@@ -1,6 +1,6 @@
-import { Devtools, liveStoreVersion } from '@livestore/common'
+import { Devtools } from '@livestore/common'
 import type { Scope } from '@livestore/utils/effect'
-import { Deferred, Effect, PubSub, Schema, Stream } from '@livestore/utils/effect'
+import { Effect } from '@livestore/utils/effect'
 
 import { makeChannelForConnectedMeshNode, makeNodeDevtoolsConnectedMeshNode } from '../web-channel/index.js'
 
@@ -38,74 +38,6 @@ export const prepareNodeDevtoolsBridge = ({
         send: Devtools.Leader.MessageToApp,
       },
     })
-
-    // const responsePubSub = yield* PubSub.unbounded<
-    //   Devtools.MessageFromApp | Devtools.MessageFromApp
-    // >().pipe(Effect.acquireRelease(PubSub.shutdown))
-
-    // // const appHostInfoDeferred = yield* Deferred.make<{ appHostId: string; isLeader: boolean }>()
-
-    // yield* nodeDevtoolsChannelCoordinator.listen.pipe(
-    //   Stream.flatten(),
-    //   // Stream.tapLogWithLabel('fromCoordinator.listen'),
-    //   Stream.tap((msg) =>
-    //     Effect.gen(function* () {
-    //       yield* PubSub.publish(responsePubSub, msg)
-    //     }),
-    //   ),
-    //   Stream.runDrain,
-    //   Effect.withSpan('portForDevtoolsChannelCoordinator.listen'),
-    //   Effect.tapCauseLogPretty,
-    //   Effect.forkScoped,
-    // )
-
-    // yield* nodeDevtoolsChannelStore.listen.pipe(
-    //   Stream.flatten(),
-    //   // Stream.tapLogWithLabel('fromStore.listen'),
-    //   Stream.tap((msg) =>
-    //     Effect.gen(function* () {
-    //       yield* PubSub.publish(responsePubSub, msg)
-    //     }),
-    //   ),
-    //   Stream.runDrain,
-    //   Effect.withSpan('portForDevtoolsChannelStore.listen'),
-    //   Effect.tapCauseLogPretty,
-    //   Effect.forkScoped,
-    // )
-
-    // // yield* nodeDevtoolsChannelCoordinator.send(Devtools.DevtoolsReady.make({ liveStoreVersion }))
-
-    // // const { appHostId, isLeader } = yield* Deferred.await(appHostInfoDeferred)
-
-    // // TODO improve disconnect handling
-    // yield* Deferred.await(nodeDevtoolsChannelCoordinator.closedDeferred).pipe(
-    //   Effect.tap(() =>
-    //     PubSub.publish(responsePubSub, Devtools.Disconnect.make({ liveStoreVersion, clientId, sessionId })),
-    //   ),
-    //   Effect.tapCauseLogPretty,
-    //   Effect.forkScoped,
-    // )
-
-    // // TODO improve disconnect handling
-    // yield* Deferred.await(nodeDevtoolsChannelStore.closedDeferred).pipe(
-    //   Effect.tap(() =>
-    //     PubSub.publish(responsePubSub, Devtools.Disconnect.make({ liveStoreVersion, clientId, sessionId })),
-    //   ),
-    //   Effect.tapCauseLogPretty,
-    //   Effect.forkScoped,
-    // )
-
-    // const sendToAppHost: Devtools.PrepareDevtoolsBridge['sendToAppHost'] = (msg) =>
-    //   Effect.gen(function* () {
-    //     // NOTE it's possible that a message is for both the coordinator and the store (e.g. Disconnect)
-    //     if (Schema.is(Devtools.MessageToApp)(msg)) {
-    //       yield* nodeDevtoolsChannelCoordinator.send(msg)
-    //     }
-
-    //     if (Schema.is(Devtools.MessageToApp)(msg)) {
-    //       yield* nodeDevtoolsChannelStore.send(msg)
-    //     }
-    //   }).pipe(Effect.withSpan('sendToAppHost'), Effect.orDie)
 
     const copyToClipboard = (text: string) =>
       Effect.sync(() => {