+++ CHANGELOG.md
@@ -32,9 +32,10 @@
     - Fix: When resetting the database but keeping the eventlog
       - the app doesn't show a shutdown screen
       - on next app start, the app doesn't re-hydrate properly (somehow seems to "double hydrate")
-    - Fix: Expo
     - Fix: Support multiple leader <> devtools connections
       - Refactor according to ARCHITECTURE.md
+- Expo:
+  - Fix:  Require cycle: ../../../packages/@livestore/livestore/dist/store/create-store.js -> ../../../packages/@livestore/livestore/dist/effect/index.js -> ../../../packages/@livestore/livestore/dist/effect/LiveStore.js -> ../../../packages/@livestore/livestore/dist/store/create-store.js
 - Tanstack Start example:
   - Fix: devtools should load smoothly on first open (currently causes a Vite error)